perf(core): shrink ElementNode to one cache line
tailwind and custom_name were both inline on every node, but only a
custom tag or a tailwind-active one ever sets them. NodeExtras boxes
the two together so a node that needs neither pays one pointer instead
of 24 bytes, and index/current_walk_index/child_text_node_index move
from usize to u32 (document positions never approach 4 billion).
Padding ElementNode back out to the old size measured 3.3-7.8% slower.

+++ crates/core/src/convert/output.rs
@@ -706,7 +706,7 @@ impl ConvertState {
     );
 
     if !self.plain_text && !enter_is_literal && tag_id == Some(TAG_LI) && !self.in_table_cell() {
-      self.record_item_marker(self.stack[stack_len - 1].index, output_start);
+      self.record_item_marker(self.stack[stack_len - 1].index as usize, output_start);
     }
 
     // The blank line that re-enables Markdown must be looked for *inside* the
@@ -2590,7 +2590,11 @@ impl ConvertState {
         s.push_str(&self.list_indent);
         if let Some(list) = ordered {
           use std::fmt::Write;
-          let _ = write!(s, "{}. ", Self::ordered_item_number(list, node.index));
+          let _ = write!(
+            s,
+            "{}. ",
+            Self::ordered_item_number(list, node.index as usize)
+          );
         } else {
           s.push_str("- ");
         }
